requirements.txt
tensorflow==2.3.0
Pillow==7.2.0